Move the main call and sys arguments management into a specific file.
"mergepdfs.py" only takes care of the command arguments management and calling the parse folder and merge files functions.
Merging the files has now moved to the specific file, "mergefiles.py".
